+++ blocks/log/__init__.py
@@ -1,0 +1,7 @@
+from .log import TrainingLog
+from .sqlite import SQLiteLog
+
+BACKENDS = {
+    'python': TrainingLog,
+    'sqlite': SQLiteLog
+}
